Refactor repository classes and interfaces
