+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/UserRepository.java
@@ -1,13 +1,4 @@
 package com.avdo.spring.app.repository;
 
-import com.avdo.spring.app.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    Optional<UserEntity> findByUsername(String username);
+public interface UserRepository {
 }
