+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/crud/CrudUserRepository.java
@@ -1,0 +1,13 @@
+package com.avdo.spring.app.repository.crud;
+
+import com.avdo.spring.app.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface CrudUserRepository extends JpaRepository<UserEntity, Long> {
+
+    Optional<UserEntity> findByUsername(String username);
+}
