+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/impl/CartItemJpaRepository.java
@@ -2,6 +2,7 @@ package com.avdo.spring.app.repository.impl;
 
 import com.avdo.spring.app.entity.CartItemEntity;
 import com.avdo.spring.app.entity.UserEntity;
+import com.avdo.spring.app.repository.CartItemRepository;
 import com.avdo.spring.app.repository.CartRepository;
 import com.avdo.spring.app.repository.crud.CrudCartItemRepository;
 import com.avdo.spring.app.repository.ProductRepository;
