+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/repository/impl/UserJpaRepository.java
@@ -1,0 +1,6 @@
+package com.avdo.spring.app.repository.impl;
+
+import com.avdo.spring.app.repository.UserRepository;
+
+public class UserJpaRepository implements UserRepository {
+}
