+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/service/CartItemServiceImpl.java
@@ -1,6 +1,6 @@
 package com.avdo.spring.app.service;
 
-import com.avdo.spring.app.repository.impl.CartItemRepository;
+import com.avdo.spring.app.repository.CartItemRepository;
 import com.avdo.spring.app.service.domain.model.CartItem;
 import com.avdo.spring.app.service.domain.request.CreateCartItemRequest;
 import org.springframework.beans.factory.annotation.Autowired;
